Squashed form '3f9174baefc3895175327e820f3f70a419b139f4' to '54385f35707241bb5fdf87aae696f94b08713c51'.
Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.

Updated the files for Docker, to see whether a local environment can be built.