Squashed from 'f5556765535f506039c2296582f63e6aebda26d9' to 'ef98f0b076dcff5ff342e5853eb570103e756e82'.
Updated 'settings.py' to fit the local and docker runtime.

Updated 'settings.py' to fit the local and docker runtime.

Updated 'settings.py' to fit the local and docker runtime.

Updated 'settings.py' to fit the local and docker runtime.

Updated 'settings.py' to fit the local and docker runtime.

Updated the configuration to fit the local and docker runtime.